requirements.txt
setuptools~=59.6.0
Twisted~=21.7.0
PyYAML~=6.0
RPi.GPIO
jsons~=1.6.1